add diff utils; add some format
