+++ page-factory-core/src/main/java/ru/sbtqa/tag/pagefactory/optional/NonCritical.java
@@ -3,8 +3,11 @@ package ru.sbtqa.tag.pagefactory.optional;
 public interface NonCritical {
 
     /**
-     * TODO
-     * @return
+     * Step is non critical, e.g.
+     * When ? user (makes some checks).
+     * '?' is a marker of a non critical step.
+     *
+     * @return true if step is non critical
      */
     boolean isNonCritical();
 }
