+++ page-factory-core/src/main/java/ru/sbtqa/tag/pagefactory/utils/GherkinResource.java
@@ -1,9 +1,7 @@
 package ru.sbtqa.tag.pagefactory.utils;
 
 import cucumber.runtime.io.Resource;
-
 import java.io.ByteArrayInputStream;
-import java.io.File;
 import java.io.InputStream;
 import java.net.URI;
 import java.nio.charset.Charset;
@@ -26,19 +24,8 @@ public class GherkinResource implements Resource {
         return path;
     }
 
-//    @Override
-//    public String getAbsolutePath() {
-//        return new File(path).getAbsolutePath();
-//    }
-
     @Override
     public InputStream getInputStream() {
         return this.source;
     }
-
-//    @Override
-//    public String getClassName(String extension) {
-//        return path.substring(0, path.length() - extension.length()).replace(File.separatorChar, '.');
-//
-//    }
 }
