+++ plugins/web-plugin/src/test/java/pagefactory/CucumberTest.java
@@ -6,7 +6,7 @@ import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(monochrome = true, plugin = {"pretty"},
-        glue = {"ru.sbtqa.tag.stepdefs",  "setting", "pagefactory.stepdefs", "ru.sbtqa.tag.pagefactory.transformer"},
+        glue = {"ru.sbtqa.tag.stepdefs",  "setting", "pagefactory.stepdefs"},
         features = {"src/test/resources/features"}
 )
 public class CucumberTest {}
