+++ pom.xml
@@ -77,7 +77,7 @@
     <module>plugins/html-plugin</module>
     <module>plugins/mobile-plugin</module>
     <module>plugins/rest-plugin</module>
-<!--    <module>plugins/mq-plugin</module>-->
+    <module>plugins/mq-plugin</module>
   </modules>
 
   <distributionManagement>
@@ -118,7 +118,6 @@
         <module>page-factory-core</module>
         <module>tests/test-web-app</module>
         <module>plugins/web-plugin</module>
-        <module>plugins/jdi-plugin</module>
         <module>plugins/html-plugin</module>
         <module>plugins/mobile-plugin</module>
         <module>plugins/rest-plugin</module>
